Fixed buggy homepage link issue
changed the placement of the RouterLink Tag,
to only the details section, so that only when its clicked,
it will redirect to details page

also merged dev and test branches to main in the prev commit.

added API testing, but needs to be careful in running it so many times,
since there is an API limit.